+++ noodles-sam/src/io/reader.rs
@@ -16,8 +16,9 @@ use noodles_bgzf as bgzf;
 use noodles_core::Region;
 use noodles_csi::BinningIndex;
 
+pub(crate) use self::record::read_record;
 pub use self::{builder::Builder, record_bufs::RecordBufs};
-use self::{header::read_header, record::read_record, record_buf::read_record_buf};
+use self::{header::read_header, record_buf::read_record_buf};
 use crate::{alignment::RecordBuf, header::ReferenceSequences, Header, Record};
 
 /// A SAM reader.